Функция spi_receive() работает корректно

+++ source/periphery/spi.h
@@ -13,6 +13,6 @@ void spi_transmit_color(const uint16_t color, const uint32_t size);
 /* Отдельно отправлять команду на чтение и читать данные нельзя
    поскольку chip select должен быть в "0" непрерывно между 
    командой и чтением данных из памяти дисплея*/
-void spi_receive(const uint8_t cmd, uint8_t *data, uint8_t size);
+void spi_receive(const uint8_t cmd, uint16_t *data, uint16_t size);
 
 #endif // __SPI_H
